broadcast: adds bc function
bc is a higher order function that accepts a procedure and a path and
returns a function that takes the same number of arguments as
procedure.

when called a udp NOTIFY broadcast will be generated with the payload
being the json representation of procedure's return value.

every rest call that results in some change that needs to be notified
to all clients should be wrapped in this function.