module: turns broadcast into a module
we should eventually turn every component in cube-server into a
module. making everything a module lets us be more explicit regarding
which symbols are exported to other files.

note that the module broadcast is still included in broadcast.scm,
this has to happen somewhere so that the runtime knows about the
modules existence.