broadcast: adds max size to udp message, ALERT verb
do avoid dealing with potential problems when udp messages become
fragmented we set a max size for udp packets

if this size is exceeded we send an ALERT message instead. As opposed
to NOTIFY messages ALERT has no body. the semantics should be
something like 'do a get request to the provided path'