refac rest player: moves functions from rest to player
when we have define-audio-handler and play-command in rest, playqueue
is dependent on rest. playqueue shouldn't know anything about rest so
moving them to player.

this reverts a previous refactoring commit where we moved the
functions from player to rest. the thinking then was to move
everything the rest-* modules depended on to its own module; rest, but
since the rest-* modules use both player as well as rest functionality
it makes sense that they require both.