restlib refac: make-search-result now converts to vector
we need vector for ->json. all users of make-search-result would convert
to vector themselves, better to have do it here only so that maybe we
can hide the json details ... ?

note that this still means that lists of lists will become array of
objects, not arrays of arrays since we convert the top-level only.