cli fix: dont run on-exit if #:quit has been invoked
since we use the on-exit callback to trigger the next song this had
unfortunate side-effects

example:

say you have a playqueue with four items; a, b, c and d. you start
playing a, on-exit will then be set to play b. Then a request to play
c arrives, we quit the process that plays a and start c. after c is
finished playing b will start playing. this happens because the
callback hangs while waiting for a mutex to be release. this also lead
to frequent crashes.