player: adds robustness in heartbeat thread
this solves two problems:

1. seek-pos-thread died on error (now it will keep going)

2. when seek-pos-thread dies, it corrupts player mutex (set it to
abandoned state) because of a bug in dynamic-wind. this no longer
appears since the thread should never die anymore.

note that the seek-post-thread dies when it tries to do
(*cplay-proc* "pos") on a process that has exited (port closed). this
should also be fixed (in process-cli).