rest api major: changes all paths to /v1/...
because automatically adding the /v1 became confusing when mixed with
NOTIFY events. better to hardcode all this until we really know what we
might want to change.

note that the define-handler macro was rewritten to check that all paths
actually start with /v1/ as a reminder. when we move to /v2/ or so,
we'll have to fix this macro, but it's been really helpful so far!!