alist-merge experimental: allow to be passed in as alist
we use alist-merge like this:

        (alist-merge (pq-current *pq*)
                     (player-pos)
                     (player-paused?))

one or more of these entries might be false, if that's the case, skip
the alist. this has the nice property of letting us return an empty
object insted of 'false' if there are no values.

NOTE: this may or may not be a good idea