bugfix critical: set current-json for current-request only
calling (current-json json) sets it "globally" for the current thread.
if we run spiffy with keep-alive connections, the current-json stays
alive for multiple requests (because it's in the same spiffy thread).

so we temporarily set current-json instead. then, on the next GET
request, it will be back to #f, and not its past POST value like without
this patch.