rest state: initial draft of rest-state
rest-state adds a new macro; define-state. the idea is that any
handler that needs to store state can define a persistent storage
location by calling (define-state foo).

This would create the route /v1/catalog/state/foo which will then
allow you to do

curl /v1/catalog/state/foo                      ;; get current state
curl /v1/catalog/state/foo -d '{"someval": 2}'  ;; update state

the system has no idea of the contents of the state so if any client
POSTs an empty json object, the state will be cleared. on each POST
every state value is written to disc.

state is persisted in a file, currently /tmp/cube-server.store, which
is just a serialized alist of all states. when initializing this file
is read back in or an empty map is created if file not found.