repl: using nrepl, removes parley/stty dependencies
because they were't working. stty can't find the terminal length on
adb shell, it goes crazy with:

```
Error: (terminal-size) Unable to get size of terminal - Success: #<output port "(stdout)">

	Call history:

	parley.scm:88: call-with-current-continuation
	parley.scm:90: with-exception-handler
	parley.scm:93: port?
	parley.scm:93: port->fileno
	parley.scm:105: char-ready?
	read-char/port
	parley.scm:285: loop
	parley.scm:284: reverse
	list->string
	parley.scm:284: string-split
	parley.scm:507: read-raw
	parley.scm:476: call-with-current-continuation
	parley.scm:478: prompt-loop
	parley.scm:419: refresh-line
	parley.scm:140: ##sys#call-with-values
	parley.scm:141: terminal-size	  	<--
```

so we decide to replace it with nrepl and avoid all stty dependencies.
that means we don't have to keep a stty fork too which is nice :)