looper: adds loop/interval function
we need this little baby should limit the frequency of calling thunk,
even if thunk throws exceptions which has been a problem in the past.

thunk can't simply do:

    (do-something-important)
    (thread-sleep! 5)

if do-something-important throws and we catch that and continue looping,
do-something-important will be ran as fast as possible and flood
everything. with loop/interval, the interval should always be respected.