fix: assign parameter rest-server-port at startup
this makes all new thread-spawns and friends use the port from the
command-line. dynamic scoping at work!

so instead of parameterizing start-rest-server, we use the existing
dynamic binding and set that before hand. start-rest-server uses this,
just like everybody else, and we have a sense of a "global variable"
without the all the evil.

this is also related to #37.