mwa: adds support for multiple wimp accounts
wimp accounts are now parameterized through a new required query
parameter; username.

When logging in we store the session in wimp-sessions with the username
as key. The client then has to refer to this username when making
queries.

The sessions are stored in memory only, so clients will need to resubmit
login credentials through a call to POST /wimp/login on login related
issues.