turi: support alists as parameter to turi creator/handler
we want to pass serveral parameters to and from turis in wimp.
Previously it was only possible to have one argument, now we have more
through the use of an alist.

The code currently contains checks to see if the argument is a string or
a list, this is there for backwards compatability. Will require the
parameter to be an alist once the dust has settled.