dab fix: don't send cplay to stream a DAB channel that isn't playing
we make the requesting zone's cplay play the returned turi.
if (dab-command (dab.sl.station cdidx)) fails, there will be no audio
payload on dab/hi, which again causes cplay to hang (ffmpeg is waiting
for 5MB of data to analyze the audio format) and not respond to any
`pause`, `quit` etc messages. this is critical because it will hang the
entire system!

before

```
[klm@kth cube-server]$ curl "linux.local:5055/v1/t2s?type=dab&id=100"
{"url":"http://127.0.0.1:8090/dab/hi"}
```

after:

```
[klm@kth cube-server]$ curl "linux.local:5055/v1/t2s?type=dab&id=100"
cannot set channel 100: ((item-set-response FS_FAIL))
((exn (location #f)
      (call-chain
        (#("restlib.scm:159: pair?" #f #f)
         #("restlib.scm:168: reverse" #f #f)
         #("restlib.scm:156: apply" #f #f)
         #("turi.scm:21: assoc" #f #f)
         #("turi.scm:20: g7244" #f #f)
         #("turi.scm:22: cdr" #f #f)
         #("turi.scm:22: g7247" #f #f)
         #("<eval>" (string->number chidxstr) #<frameinfo>)
         #("<eval>" (pp `(rest-dab station ,chidx)) #<frameinfo>)
         #("<eval>"
           (##sys#list (##core#quote rest-dab) (##core#quote station) chidx)
           #<frameinfo>)
         #("<eval>" (dab-command (dab.sl.station chidx)) #<frameinfo>)
         #("<eval>" (dab.sl.station chidx) #<frameinfo>)
         #("<eval>"
           (equal?1286 v1265 (quote1284 (item-set-response FS_OK)))
           #<frameinfo>)
         #("<eval>" (failure1271) #<frameinfo>)
         #("<eval>"
           (error (conc "cannot set channel " chidxstr) anything)
           #<frameinfo>)
         #("<eval>" (conc "cannot set channel " chidxstr) #<frameinfo>)))
      (arguments ((item-set-response FS_FAIL)))
      (message "cannot set channel 100")))
```

this fixes a problem with old favorit channels stored on the phones.
those turis are not "permanent" because they store the DAB channel
index, which may change on every dab scan.

Fises part 1 of #60.