dns txt: makes service names pseudo unique
using publish name as id for the this instance. Name shown to users is
now stored in name field of txt record json.

This is very much a hack to get around the fact that we're killing the
service and re-publishing it when the txt-record changes. Gets us around
issues of the old service lingers around after a new one has been published.

The proper solution would probably be to use the avahi library directly
instead of using avahi-publish. Than we could update the record without
starting a new service.