fm refac: only run notify when we're searching
the poll based notify thread is expensive and we only really need it
when we're searching to inform clients of the progress.

only downside to this is that clients won't be notificed if radio-text
suddenly becomes available. it's rarely non-empty anyway because of
lousy signal strength. we'll see once we get a proper antenna.