notify bugfix: fixes incoming requests hanging
running

    while true ; do curl localhost:5055/v1/player/notify & sleep .1; kill $! ; done

for a while would eventually have all cube-server requests hang. this is
because of
http://bugs.call-cc.org/browser/release/4/spiffy/trunk/spiffy.scm#L617
that doesn't decrement thread-count and makes Spiffy think there are
1024 simultanious connections when there really are only a couple.

this fixes the problem but introduces an ugly error message that happens
every time a client disconnects from `/v1/[player|catalog]/notify`. we
can fix that later.

fixes #75