dab/fm: adds a lock during dab scanning
we only do dab scanning once; during startup. if the user is quick you
can cancel the scanning process by starting fm. when this happens
a reboot is needed to get dab working.

this patch adds a guard during scanning that returns 503 Service
Unavailable while the initial dab scan is happening.

also adds a route in the dummy api to emulate this. By calling
/v1/catalog/dab/fail the server will 'fail' for the next 30 seconds