dab refac: move general dab utils dab-i2c => dab
they don't have anything to do with i2c. however, they need some
dab-command interface. maybe we should have made a functor here but it
seems like so much work for just a very simple API.

let's see if this works first (but the define dab-command with error is ugly!)