dab bugfix: set dab-scanning flag correctly on dab errors
the (set! ..) after dab-refresh-channels will never be called if there's
a dab error. this will lock the frontend with "DAB is scanning..."
forever and block all sorts of API calls.

we still want to retry after a while on errors, because there may be
FSNODE_BLOCKED errors or other obscurities where we'd want to retry. but
not every 10 seconds, retry every minute.