/v1/version: workaround cross-compilation git version
because `git describe` will fail when building for target, because it
copies cube-server/* and not cube-server/[.]* so .git is missing.

I tried to cache version to a string too, but that didn't work. I think
the correct way is either:

- to have chicken-install set an
environment-variable (target-compilation feature) that let's you access
the original directory.

- to copy hidden files too. I think this is easier and cleaner.

see #58