issue #105: wrong stations playing
when reading DAB channels using dab.sl.uComponent we didn't
store the index we were quering. Instead we stored the id returned in
field #x02; serviceKey.

According to the fsapi documentation the query argument to uComponent is
called componentKey. ComponentKey and serviceKey are usually, but not
always, the same thing.

This patch stores the actual index we queried instead of the serviceKey
in the response.