minor fix: don't include raw request in current
the raw json-request can contain all kinds of junk from the clients.

previously we included data from the client _over_ fields the server
made itself. this had the unfortunate effect of removing the id from
current if the client encoded a missing id as an empty string (which the
android client does)