dlna: parse duration meta-information from DLNA server
before patch, no duration field:

$ curl "http://192.168.0.182:5055/v1/catalog/dlna/browse?id=b882528c2.1%246%248%241&q=&limit=1&offset=0" | jq
{"limit":1,"offset":0,"total":23,"items":[{"turi":"http://192.168.0.182:8200/MediaItems/203.mp3","title":"Adagio","subtitle":"Lara Fabian","album":"Lara Fabian"}]}

after patch, duration field (in seconds) pops up:

$ curl -s "http://192.168.0.182:5055/v1/catalog/dlna/browse?id=b882528c2.1%246%248%241&q=&limit=1&offset=0" | jq .
{
  "limit": 1,
  "offset": 0,
  "total": 23,
  "items": [
    {
      "turi": "http://192.168.0.182:8200/MediaItems/203.mp3",
      "title": "Adagio",
      "subtitle": "Lara Fabian",
      "album": "Lara Fabian",
      "duration": 267.758
    }
  ]
}

this is necessary because clients (iCube) will assume it's a live stream
if the duration field is not present. we should be more explicit about
this (duration : false for live streams, for example), but this will at
least fix all tracks from DLNA poping up as "live streams" for now.