fix: don't store pos value in playqueue
if a client is transferring a song from one speaker to another we want
that song to begin at roughly the same place when playback on the new
speaker begins. If this position is stored in the playqueue the song
will start at that point when it's selected from the playqueue as well,
clearly not what we want.