test: update turi tests to reflect new behaviour
we used to have a hack (#93) that would always set host 127.0.0.1 so
that DAB radio favorites would work. now with solo2 speakers, this will
no longer work: solo2 speakers need to know the ip/hostname of the
catalog so they can in the tr:// turis. this was fixed long
ago (removing the #93 hack), breaking these tests.

now these tests pass, but the #93 problem still applies: DAB favorites
still contain an IP address in the DAB favorites turi. that's a TODO.