wimp: adds enpoints for generated playlists
add endpoints for the following generated playlists:
 - featured/new
 - discover
 - genres
 - moods

the abstractions we're using for specifying rest-apis is starting to
show some tearing with the introduction of these routes. especially the
introduction of GET routes that do not contain a query parameter led to
a lot of code duplication.

fixes jb/platform_manifest#14