feature wimp: preload routes for user favorites
Adds support for TIDAL favorites. Unfortunately we have to use a custom
handler for these endpoints since the objects returned use a different
format from the others

The favorites endpoints returns an object that looks like this:

{
  "limit": 10,
  "offset": 0,
  "totalNumberOfItems": 2,
  "items": [
    {
      "created": "2016-07-06T12:38:40.815+0000",
      "item": {
        "id": 36934963,
        "title": "Megalithic Symphony Deluxe",
....

Other enpoints return this:

{
  "limit": 10,
  "offset": 0,
  "totalNumberOfItems": 62,
  "items": [
    {
      "id": 62132141,
      "title": "Earth",
....

Notice that each item in favorites has two keys; 'created' and 'item',
instead of the contents of 'item' as the object itself.

We handle this difference by mapping and extracting the value under
'item', ie:

  (map (lambda (x) (alist-ref 'item x)) (alist-ref 'items result))

this implements jb/platform_manifest#119