fix pos: illegal comparison
the comparision assumes that `duration` is present in json-request, this
turns out to be wrong most of the time. this leds to a server error when
we compare against #f

check that we have a current song (pos changes don't make much sense
without one) and compare against current songs duration.